Fix incorrect dependency on xds_url_map_testcase for CSDS logic
