+++ framework/rpc/grpc_csds.py
@@ -15,9 +15,12 @@
 This contains helpers for gRPC services defined in
 https://github.com/envoyproxy/envoy/blob/main/api/envoy/service/status/v3/csds.proto
 """
-
+import json
 import logging
-from typing import Optional
+import re
+from typing import Any, Optional
+
+from typing_extensions import TypeAlias
 
 # Needed to load the descriptors so that Any is parsed
 # TODO(sergiitk): replace with import xds_protos when it works
@@ -37,8 +40,100 @@ import framework.rpc
 logger = logging.getLogger(__name__)
 
 # Type aliases
-ClientConfig = csds_pb2.ClientConfig
-_ClientStatusRequest = csds_pb2.ClientStatusRequest
+ClientConfig: TypeAlias = csds_pb2.ClientConfig
+_ClientStatusRequest: TypeAlias = csds_pb2.ClientStatusRequest
+ClientStatusResponse: TypeAlias = csds_pb2.ClientStatusResponse
+JsonType: TypeAlias = dict[Any]
+
+
+class DumpedXdsConfig(dict):
+    """
+    A convenience class to check xDS config.
+
+    Feel free to add more pre-compute fields.
+    """
+
+    def __init__(self, xds_json: JsonType):  # pylint: disable=too-many-branches
+        super().__init__(xds_json)
+        self.json_config = xds_json
+        self.lds = None
+        self.rds = None
+        self.rds_version = None
+        self.cds = []
+        self.eds = []
+        self.endpoints = []
+        for xds_config in self.get("xdsConfig", []):
+            try:
+                if "listenerConfig" in xds_config:
+                    self.lds = xds_config["listenerConfig"]["dynamicListeners"][
+                        0
+                    ]["activeState"]["listener"]
+                elif "routeConfig" in xds_config:
+                    self.rds = xds_config["routeConfig"]["dynamicRouteConfigs"][
+                        0
+                    ]["routeConfig"]
+                    self.rds_version = xds_config["routeConfig"][
+                        "dynamicRouteConfigs"
+                    ][0]["versionInfo"]
+                elif "clusterConfig" in xds_config:
+                    for cluster in xds_config["clusterConfig"][
+                        "dynamicActiveClusters"
+                    ]:
+                        self.cds.append(cluster["cluster"])
+                elif "endpointConfig" in xds_config:
+                    for endpoint in xds_config["endpointConfig"][
+                        "dynamicEndpointConfigs"
+                    ]:
+                        self.eds.append(endpoint["endpointConfig"])
+            # TODO(lidiz) reduce the catch to LookupError
+            except Exception as e:  # pylint: disable=broad-except
+                logging.debug(
+                    "Parsing dumped xDS config failed with %s: %s", type(e), e
+                )
+        for generic_xds_config in self.get("genericXdsConfigs", []):
+            try:
+                if re.search(r"\.Listener$", generic_xds_config["typeUrl"]):
+                    self.lds = generic_xds_config["xdsConfig"]
+                elif re.search(
+                    r"\.RouteConfiguration$", generic_xds_config["typeUrl"]
+                ):
+                    self.rds = generic_xds_config["xdsConfig"]
+                    self.rds_version = generic_xds_config["versionInfo"]
+                elif re.search(r"\.Cluster$", generic_xds_config["typeUrl"]):
+                    self.cds.append(generic_xds_config["xdsConfig"])
+                elif re.search(
+                    r"\.ClusterLoadAssignment$", generic_xds_config["typeUrl"]
+                ):
+                    self.eds.append(generic_xds_config["xdsConfig"])
+            # TODO(lidiz) reduce the catch to LookupError
+            except Exception as e:  # pylint: disable=broad-except
+                logging.debug(
+                    "Parsing dumped xDS config failed with %s: %s", type(e), e
+                )
+        for endpoint_config in self.eds:
+            for endpoint in endpoint_config.get("endpoints", {}):
+                for lb_endpoint in endpoint.get("lbEndpoints", {}):
+                    try:
+                        if lb_endpoint["healthStatus"] == "HEALTHY":
+                            self.endpoints.append(
+                                "%s:%s"
+                                % (
+                                    lb_endpoint["endpoint"]["address"][
+                                        "socketAddress"
+                                    ]["address"],
+                                    lb_endpoint["endpoint"]["address"][
+                                        "socketAddress"
+                                    ]["portValue"],
+                                )
+                            )
+                    # TODO(lidiz) reduce the catch to LookupError
+                    except Exception as e:  # pylint: disable=broad-except
+                        logging.debug(
+                            "Parse endpoint failed with %s: %s", type(e), e
+                        )
+
+    def __str__(self) -> str:
+        return json.dumps(self, indent=2)
 
 
 class CsdsClient(framework.rpc.grpc.GrpcClientHelper):
